backend/fastapi-api/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
python-multipart>=0.0.6
pydantic>=2.3.0
email-validator>=2.0.0
aiofiles>=23.1.0
Pillow>=10.0.0
mutagen>=1.47.0
httpx >= 0.24.1